Add 3D Files, accept callback if data send.
Add 3D Files for the left Button for customization.
Also provide a callback if data has been transmitted to a device. The device has to write to the readCharacteristic. This way the monitor know that the measurement has been transmitted. Otherwise new notifications are send. This provides the option for a single measurement to be transmitted later.